# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CodeMind API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/admin/profile-reviews/[userId]/reject": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "API endpoint for /api/admin/profile-reviews/[userId]/reject",
        "description": "Tự động sinh tài liệu cho POST /api/admin/profile-reviews/[userId]/reject. Hãy cập nhật mô tả chi tiết sau.",
        "responses": {
          "200": {
            "description": "Thành công"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        }
      }
    },
...
    "/api/admin/users/[userId]/verifications": {
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "API endpoint for /api/admin/users/[userId]/verifications",
        "description": "Tự động sinh tài liệu cho PUT /api/admin/users/[userId]/verifications. Hãy cập nhật mô tả chi tiết sau.",
        "responses": {
          "200": {
            "description": "Thành công"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        }
      }
    },
...
    "/api/ai/complete": {
      "post": {
        "tags": [
          "Ai"
        ],
        "summary": "API endpoint for /api/ai/complete",
        "description": "Tự động sinh tài liệu cho POST /api/ai/complete. Hãy cập nhật mô tả chi tiết sau.",
        "responses": {
          "200": {
            "description": "Thành công"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        }
      }
    },
...
    "/api/ai/generate-code": {
      "post": {
        "tags": [
          "Ai"
        ],
        "summary": "API endpoint for /api/ai/generate-code",
        "description": "Tự động sinh tài liệu cho POST /api/ai/generate-code. Hãy cập nhật mô tả chi tiết sau.",
        "responses": {
          "200": {
            "description": "Thành công"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        }
      }
    },
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "API endpoint for /api/auth/login",
        "description": "Tự động sinh tài liệu cho POST /api/auth/login. Hãy cập nhật mô tả chi tiết sau.",
        "responses": {
          "200": {
            "description": "Thành công"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        }
      }
    },
...
  },
...
}